pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codi"
version = "0.1.0"
description = "Modern Discord bot with pluggable personalities (CODI for crypto, Wooloo for Pokemon/AI)."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Eric Hare", email = "ericrhare@gmail.com" }]
dependencies = [
    "discord.py>=2.4.0",
    "httpx>=0.27.0",
    "anthropic>=0.39.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.1",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[project.scripts]
codi = "codi.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/codi"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "ASYNC", "S", "SIM", "RET"]
ignore = ["S101"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
